Add list of transactions in Addresses screen
Closes #79
